* VCompartment (caseSubjectMembership): Moved from VRequirement to apply it to analyisCases. * VRequirement (caseSubjectMembership): Moved to VCompartment.
